// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Plan = {
  FREE: 'FREE',
  PRO: 'PRO'
} as const

export type Plan = (typeof Plan)[keyof typeof Plan]


export const Conversion_Status = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type Conversion_Status = (typeof Conversion_Status)[keyof typeof Conversion_Status]